# Store: improve index header reading performance by sorting values