<!--
    Keep PR title verbose enough and add prefix telling
    about what components it touches e.g "query:" or ".*:"
-->

Users may construct a promQL like `metric{key=~"b|z|..tens of thousands..|a"}`  with Grafana variables or code, which is extremely slow.

In our environment, the index header is ~99MiB per block, the CPU usage has doubled immediately when the store gateway receives a query like this(which contains ~15000 values separated by |)..

I have seen Prometheus already does something like this: https://github.com/prometheus/prometheus/blob/bcd548c88b06543c8eeb19e68bef4adefb7b95fb/tsdb/querier.go#L321

<!--
    Don't forget about CHANGELOG!

    Changelog entry format:
    - [#<PR-id>](<PR-URL>) Thanos <Component> ...

    <PR-id> Id of your pull request.
    <PR-URL> URL of your PR such as https://github.com/thanos-io/thanos/pull/<PR-id>
    <Component> Component affected by your changes such as Query, Store, Receive.
-->

* [x] I added CHANGELOG entry for this change.
* [ ] Change is not relevant to the end user.

## Changes

<!-- Enumerate changes you made -->

## Verification

<!-- How you tested it? How do you know it works? -->