@GiedriusS 2022-08-12
Benchmark diff:

```
name                                                           old time/op    new time/op    delta
BucketSeries/1000000SeriesWith1Samples/1of1000000-16             77.5ms ± 7%    79.5ms ± 9%    ~     (p=0.081 n=15+15)
BucketSeries/1000000SeriesWith1Samples/10of1000000-16            75.9ms ± 9%    82.1ms ± 6%  +8.18%  (p=0.000 n=15+15)
BucketSeries/1000000SeriesWith1Samples/1000000of1000000-16        775ms ± 3%     774ms ± 5%    ~     (p=0.949 n=14+15)
BucketSeries/100000SeriesWith100Samples/1of10000000-16           5.36ms ± 3%    5.31ms ± 2%    ~     (p=0.089 n=15+15)
BucketSeries/100000SeriesWith100Samples/100of10000000-16         5.30ms ± 3%    5.31ms ± 2%    ~     (p=0.595 n=15+15)
BucketSeries/100000SeriesWith100Samples/10000000of10000000-16    76.6ms ± 7%    74.5ms ± 3%  -2.67%  (p=0.045 n=15+15)
BucketSeries/1SeriesWith10000000Samples/1of10000000-16            126µs ± 3%     124µs ± 2%  -1.36%  (p=0.001 n=13+14)
BucketSeries/1SeriesWith10000000Samples/100of10000000-16          126µs ± 1%     124µs ± 2%  -1.87%  (p=0.000 n=14+15)
BucketSeries/1SeriesWith10000000Samples/10000000of10000000-16    15.8ms ± 2%    14.9ms ± 6%  -5.51%  (p=0.000 n=13+14)

name                                                           old alloc/op   new alloc/op   delta
BucketSeries/1000000SeriesWith1Samples/1of1000000-16             60.9MB ± 0%    60.9MB ± 0%    ~     (p=0.201 n=15+14)
BucketSeries/1000000SeriesWith1Samples/10of1000000-16            60.9MB ± 0%    60.9MB ± 0%    ~     (p=0.285 n=15+15)
BucketSeries/1000000SeriesWith1Samples/1000000of1000000-16       1.27GB ± 0%    1.27GB ± 0%    ~     (p=0.583 n=13+14)
BucketSeries/100000SeriesWith100Samples/1of10000000-16           4.61MB ± 0%    4.61MB ± 0%    ~     (p=0.715 n=14+15)
BucketSeries/100000SeriesWith100Samples/100of10000000-16         4.62MB ± 0%    4.61MB ± 0%    ~     (p=0.139 n=12+15)
BucketSeries/100000SeriesWith100Samples/10000000of10000000-16     116MB ± 5%     117MB ± 4%    ~     (p=0.377 n=15+14)
BucketSeries/1SeriesWith10000000Samples/1of10000000-16            198kB ± 0%     198kB ± 0%    ~     (p=0.862 n=15+15)
BucketSeries/1SeriesWith10000000Samples/100of10000000-16          198kB ± 0%     198kB ± 0%    ~     (p=0.461 n=15+15)
BucketSeries/1SeriesWith10000000Samples/10000000of10000000-16    38.1MB ± 0%    38.1MB ± 0%    ~     (p=0.539 n=15+15)

name                                                           old allocs/op  new allocs/op  delta
BucketSeries/1000000SeriesWith1Samples/1of1000000-16              9.79k ± 0%     9.79k ± 0%    ~     (p=0.457 n=15+14)
BucketSeries/1000000SeriesWith1Samples/10of1000000-16             9.89k ± 0%     9.90k ± 0%    ~     (p=0.229 n=15+15)
BucketSeries/1000000SeriesWith1Samples/1000000of1000000-16        10.0M ± 0%     10.0M ± 0%    ~     (p=0.675 n=15+15)
BucketSeries/100000SeriesWith100Samples/1of10000000-16            1.19k ± 0%     1.19k ± 0%    ~     (p=0.565 n=15+15)
BucketSeries/100000SeriesWith100Samples/100of10000000-16          1.23k ± 0%     1.23k ± 0%  -0.12%  (p=0.011 n=15+15)
BucketSeries/100000SeriesWith100Samples/10000000of10000000-16     1.00M ± 0%     1.00M ± 0%    ~     (p=0.262 n=15+15)
BucketSeries/1SeriesWith10000000Samples/1of10000000-16              280 ± 0%       280 ± 0%    ~     (all equal)
BucketSeries/1SeriesWith10000000Samples/100of10000000-16            280 ± 0%       280 ± 0%    ~     (all equal)
BucketSeries/1SeriesWith10000000Samples/10000000of10000000-16      168k ± 0%      168k ± 0%  -0.00%  (p=0.000 n=15+15)
```

One has even increased :scream: do you get similar results? It seems like there isn't much of an improvement :thinking: I haven't had enough time to look deeply into this. Perhaps you know the reason?