@damnever 2022-08-13
Try this: https://gist.github.com/damnever/b5bf73877185dc24fa087792cc7b284a

```
goos: darwin
goarch: amd64
pkg: github.com/thanos-io/thanos/pkg/store
cpu: Intel(R) Core(TM) i7-6700HQ CPU @ 2.60GHz

name                                                                     old time/op    new time/op    delta
BucketIndexReader_ExpandedPostings2/n=~"101|random-shuffled-values|1"-8    32.1ms ± 5%    23.7ms ± 4%  -26.09%  (p=0.000 n=9+9)

name                                                                     old alloc/op   new alloc/op   delta
BucketIndexReader_ExpandedPostings2/n=~"101|random-shuffled-values|1"-8    9.13MB ± 0%    7.08MB ± 0%  -22.37%  (p=0.000 n=10+9)

name                                                                     old allocs/op  new allocs/op  delta
BucketIndexReader_ExpandedPostings2/n=~"101|random-shuffled-values|1"-8      170k ± 0%      120k ± 0%  -29.40%  (p=0.000 n=10+10)
```

I'd expect a much better result on HDD.