@bwplotka 2022-08-16
Thanks! I think it makes sense to have this merged. The latency spike is mostly on small (10 values), so it's fine.

But I believe we would love your benchmark to have in, so any other developer can verify and we can also ensure next changes maintains the same level of efficiency 🤗 

WDYT @damnever ?